Mini_Sheh v0.2 : Ajout de l'environnement sous forme de liste chaînée, ajout de 3 built-ins supplémentaires, modification de quelques fonctions de la libft pour une meilleure compatibilité

+++ Sources/libft/ft_strlen.c
@@ -3,18 +3,18 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strlen.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: hcremers <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: hcremers <hcremers@student.s19.be>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/04 13:36:15 by hcremers          #+#    #+#             */
-/*   Updated: 2021/10/04 13:36:18 by hcremers         ###   ########.fr       */
+/*   Updated: 2022/05/03 13:57:04 by hcremers         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t	ft_strlen(const char *s)
+int	ft_strlen(const char *s)
 {
-	size_t	count;
+	int	count;
 
 	count = 0;
 	while (s[count])
